clippy fix

+++ vegafusion-core/src/data/table.rs
@@ -4,7 +4,7 @@ use crate::arrow::{
     record_batch::RecordBatch,
 };
 use crate::error::{Result, ResultWithContext, VegaFusionError};
-use std::collections::HashSet;
+
 use std::convert::TryFrom;
 use std::sync::Arc;
 
@@ -16,7 +16,7 @@ use std::io::Cursor;
 use super::scalar::ScalarValue;
 use crate::arrow::array::ArrayRef;
 use crate::data::json_writer::record_batches_to_json_rows;
-use crate::data::scalar::ScalarValueHelpers;
+
 use arrow::array::{Date32Array, Int64Array, StructArray};
 use arrow::compute::{cast, unary};
 use arrow::datatypes::TimeUnit;
@@ -143,7 +143,7 @@ impl VegaFusionTable {
                         cast(col, &DataType::Int64).unwrap()
                     }
                     DataType::Date32 => {
-                        let ms_per_day = 1000 * 60 * 60 * 24 as i64;
+                        let ms_per_day = 1000 * 60 * 60 * 24_i64;
                         let array = col.as_any().downcast_ref::<Date32Array>().unwrap();
 
                         let array: Int64Array = unary(array, |v| (v as i64) * ms_per_day);
